app/api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyYAML==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
gunicorn==21.2.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
